Morosos
selecciona bien pero muestra fecha loca, podemos sacar la fecha del listado

+++ app/src/main/java/com/appminds/clubdeportivo/data/db/contracts/ClientContract.kt
@@ -28,7 +28,7 @@ object ClientContract {
             ${Columns.TELEFONO} TEXT NOT NULL,
             ${Columns.DOMICILIO} TEXT NOT NULL,
             ${Columns.FECHA_ALTA} TEXT NOT NULL,
-            ${Columns.FECHA_VENCIMIENTO} INTEGER NOT NULL DEFAULT 0,
+            ${Columns.FECHA_VENCIMIENTO} LONG NOT NULL DEFAULT 0,
             ${Columns.APTO_FISICO} INTEGER NOT NULL DEFAULT 0,
             ${Columns.TIPO_CLIENTE} TEXT NOT NULL,
             ${Columns.ESTADO} TEXT NOT NULL
